Add -mesh and -flowhelp options.
This changes basic functionality, but the new version is documented both
in the README.md and in doc/stokes.pdf.  The added option -mesh is the
same as the required positional argument for the mesh file name.  The
added option -flowhelp is the same as the old -h.  Now -flowhelp works
for flow.py options, while -help works for PETSc options.